I have adjusted some stuff
I adjusted how "peeps" will display in the adjacent column. The user inputs will now be a different colour and font size (h5) in order to make it easier to differentiate from the subheaders.

I have also tried to adjust the media queries for phones with thinner and longer screens.