Added some localstorage stuff
BRUh. I am tired of this already. I added a function "storeIGN" that is called at the end of the function that activates once you hit the yupButton. That's fun. And Neat.

Now I guess I'll have to create a button that loads all previous entries. I tried to make it an onload function but got an error because apparently I'm loading something too early.

Fix it.