I added a random ID Generator
It works. Now I need to get a loop functioning. This loop should iterate through each "peep" that is saved in local storage, and for each peep it finds, and create a card for it. This will be the "load older peeps" button.

You left some comments for yourself. Please read them carefully and think about what you're going to do before you do it. TT_____TT